4. To infinity and beyond

+++ 0x0A-argc_argv/4-add.c
@@ -9,7 +9,7 @@
 
 int main(int argc, char *argv[])
 {
-	int i = 1, addi_res = 0;
+	long int i = 1, addi_res = 0;
 
 	if (argc > 1)
 	{
